Add files via upload
Let me present to you the Age Calculator – a very basic, yet a comfortable utility that will help you calculate your age in no time. The attractive interface and large buttons make it easy to know what age one is in years, months, or sometimes even days with ease. Are you interested in knowing when your birthday is or do you want to have fun with the time? Use this delightful calculator!